LayoutTests/inspector/canvas/console-record-offscreen-canvas-2d-expected.txt

Test that CanvasManager is able to record actions made to offscreen 2D canvas contexts.


== Running test suite: Canvas.recording2D
-- Running test case: Canvas.recording2D.Console.OffscreenCanvas2D
PASS: The recording should have the name "TEST".
PASS: The recording should have one frame.
PASS: The first frame should have one action.
